\name{PBSddesolve}
\alias{PBSddesolve-package}
\alias{PBSddesolve}
\docType{package}
\title{
  Package: Solver for Delay Differential Equations
}
\description{
  A solver for systems of delay differential equations based on 
  numerical routines from 
  \href{https://www.maths.ed.ac.uk/~swood34/simon/dde.html}{Simon Wood}'s
  \code{solv95} program. 
  This solver is also capable of solving systems of ordinary
  differential equations.
}
\details{
  Please see the user guide \code{PBSddesolve-UG.pdf}, located in R's
  library directory \code{./library/PBSddesolve/doc}, for a 
  comprehensive overview.
}
\author{
  Alex Couture-Beil <alex@mofo.ca> \cr
  Jon T. Schnute <schnutej@shaw.ca> \cr
  Rowan Haigh <rowan.haigh@dfo-mpo.gc.ca>\cr
  Nicholas Boers <nboers@gmail.com>

  Maintainer: \href{mailto:rowan.haigh@dfo-mpo.gc.ca}{Rowan Haigh}, Program Head -- Offshore Rockfish\cr
  Pacific Biological Station (PBS), Fisheries & Oceans Canada (DFO), Nanaimo BC\cr
  \emph{locus opus}: Regional Headquarters (RHQ), Vancouver BC\cr
  Last modified \code{Rd: 2022-10-25}
}
\references{
  Wood, S.N. (1999) Solv95: a numerical solver for systems of delay 
  differential equations with switches. Saint Andrews, UK. 10 pp.
}

\seealso{
  \code{\link[PBSddesolve]{dde}}
}
\keyword{package}

