\name{PBSddesolve}
\alias{PBSddesolve-package}
\alias{PBSddesolve}
\docType{package}
\title{
Package: Solver for Delay Differential Equations
}
\description{
A solver for systems of delay differential equations based off 
numerical routines from Simon Wood's \code{solv95} program. 
This solver is also capable of solving systems of ordinary differential equations.
}
\details{
Please see the user guide \code{PBSddesolve-UG.pdf}, located in \code{.../library/PBSddesolve/doc},
 for a comprehensive overview.
}

\author{
  Alex Couture-Beil <alex\_rproject@mofo.ca>\cr
  Jon T. Schnute <Jon.Schnute@dfo-mpo.gc.ca>\cr
  Rowan Haigh <Rowan.Haigh@dfo-mpo.gc.ca>

Maintainer: Jon Schnute <Jon.Schnute@dfo-mpo.gc.ca>
}

\references{
Wood, S.N. (1999) Solv95: a numerical solver for systems of delay 
differential equations with switches. Saint Andrews, UK. 10 pp.
URL: \url{http://www.maths.bath.ac.uk/~sw283/simon/dde.html}
}

\seealso{ \code{\link{dde}} }

\keyword{ package }

